refact: change concurrency to 1
rename deleteMessage to enqueueDeleteMessage
fix up some style stuff

+++ index.js
@@ -4,42 +4,42 @@
 // @version    1.0
 // ==/UserScript==
 
-(function() {
-  'use strict';
-
-  function makeRequest (url, params) {
-    return new Promise(function (resolve, reject) {
-      var xhr = new XMLHttpRequest();
-      xhr.open('POST', url, true);
+;(function() {
+  'use strict'
+  function makeRequest(url, params) {
+    return new Promise(function(resolve, reject) {
+      var xhr = new XMLHttpRequest()
+      xhr.open('POST', url, true)
       xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded')
-      xhr.onload = function () {
+      xhr.onload = function() {
         if (this.status >= 200 && this.status < 300) {
-          resolve(xhr.response);
+          resolve(xhr.response)
         } else {
           reject({
             status: this.status,
-            statusText: xhr.statusText
-          });
+            statusText: xhr.statusText,
+          })
         }
-      };
-      xhr.onerror = function () {
+      }
+      xhr.onerror = function() {
         reject({
           status: this.status,
-          statusText: xhr.statusText
-        });
-      };
-      xhr.send(params);
-    });
+          statusText: xhr.statusText,
+        })
+      }
+      xhr.send(params)
+    })
   }
 
-  const _x_id = () => `${window.boot_data.version_uid.substring(0, 8)}-${(new Date).getTime()/1000}`
+  const _x_id = () =>
+    `${window.boot_data.version_uid.substring(0, 8)}-${new Date().getTime() / 1000}`
 
-  var maxConcurrent = 3;
-  var maxQueue = Infinity;
-  var queue = new Queue(maxConcurrent, maxQueue);
+  var maxConcurrent = 1
+  var maxQueue = Infinity
+  var queue = new Queue(maxConcurrent, maxQueue)
 
   const queryString = obj => {
-    const str = [];
+    const str = []
     for (const p in obj)
       if (obj.hasOwnProperty(p)) {
         str.push(encodeURIComponent(p) + '=' + encodeURIComponent(obj[p]))
@@ -47,7 +47,7 @@
     return str.join('&')
   }
 
-  const deleteMessage = (...args) => 
+  const enqueueDeleteMessage = (...args) =>
     queue.add(() =>
       (({ channel, ts }) => {
         const token = window.boot_data.api_token
@@ -62,15 +62,17 @@
     if (!el) return
     const ts = el.getAttribute('data-ts')
     const channel = el.getAttribute('data-model-ob-id')
-    deleteMessage({ channel, ts })
+    enqueueDeleteMessage({ channel, ts })
     el.parentNode.removeChild(el)
   }
 
+  const shouldDelete = e => e.shiftKey && e.keyCode === 8 // shift + backspace
+
+  // const shouldDelete = e => e.ctrlKey && e.shiftKey && e.code === 'Space'
+
   const onKeyUp = e => {
-    if (e.ctrlKey && e.shiftKey && e.code === 'Space') {
-      deleteHoveredMessage()
-    }
+    if (shouldDelete(e)) deleteHoveredMessage()
   }
 
   document.addEventListener('keyup', onKeyUp, false)
-})();
+})()
